refactor(auth): merge and centralize types and schemas for signup and institution
