+++ src/features/auth/components/InstitutionForm.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { InstitutionPayload } from '@features/auth/types';
-import { institutionSchema } from '@features/auth/schemas';
+import { InstitutionPayload } from '@features/auth/type';
+import { institutionSchema } from '@features/auth/schema';
 import { institutionService } from '@api/auth/institution';
 import { ApiError, INSTITUTION_ERROR_MAP, UNKNOWN_ERROR } from '@custom-types/ApiError';
 import { getErrorField } from '@utils/errorHandler';
