+++ src/features/auth/type.ts
@@ -34,3 +34,31 @@ export interface LoginFormData  {
   email: string;
   password: string;
 };
+
+export type SignUpInput = {
+  firstName: string;
+  lastName: string;
+  username: string;
+  email: string;
+  code: string;
+  password: string;
+  confirmPassword: string;
+  agreeTerms: boolean;
+};
+
+export type InstitutionPayload = {
+  name: string;
+  slug: string;
+  emailDomain: string;
+  logo?: File;
+};
+
+export type SignupFormData = {
+  firstname: string;
+  lastname: string;
+  username: string;
+  email: string;
+  password: string;
+  confirmPassword?: string;
+  verifyCode: string;
+};
